tests: Do not depend on useradd in TEST-19-CGROUP.delegate
The current form allows failure of useradd but then runs subsequent test
anyway. Bail out and report a skipped (sub)test.

+++ test/units/TEST-19-CGROUP.delegate.sh
@@ -74,7 +74,10 @@ testcase_attributes() {
 
 testcase_scope_unpriv_delegation() {
     # Check that unprivileged delegation works for scopes
-    useradd test ||:
+    if ! useradd test ; then
+        echo "Skipping $0 unprivileged delegation test, can't create users" >&2
+        return
+    fi
     trap "userdel -r test" RETURN
     systemd-run --uid=test \
                 --property="User=test" \
